vms-ml-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
pymysql==1.1.1
python-dotenv==1.0.1
pandas==2.2.3
numpy==2.1.1
scipy==1.14.1
scikit-learn==1.5.2
statsmodels==0.14.4
python-jose[cryptography]==3.3.0
httpx==0.27.2
